+++ release_notes.md
@@ -2,6 +2,8 @@
 
 ## NEXT
 * [TAB-29182] Optima Digital Signature
+* [TAB-33620] DCV | AU Modifications for Gift Card Load
+* [TAB-33679] DCV | AU | Add the word refund before the payment if the payment type is refund
 
 ## 1.54.1
 * [TAB-33702] DCV | Regression 7.21.1 - Damaged document viewer